pyproject.toml
[project]
name = "pynf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jpype1>=1.6.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
setup-nextflow = "setup_nextflow:main"
pynf = "pynf.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["src/pynf"]

[tool.hatch.build.targets.wheel]
packages = ["src/pynf"]
